requirements.txt
# Flocalize Project Dependencies
# AI and LLM Libraries
google-genai==1.30.0
anthropic==0.64.0
xai-sdk==1.0.1

# Environment and Configuration
python-dotenv==1.1.1

# Census Data Analysis
census>=0.8.24   # https://pypi.org/project/census/

# Data Analysis and Visualization (for census examples)
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0

# Interactive Mapping Libraries
plotly==6.3.0     # Interactive plotting and maps
kaleido>=1.0.0    # Image export for Plotly

# US Geographic Data Helper
us>=3.1.1

# Web Framework and Utilities
requests>=2.32.0
